store scan results in storage
