+++ httphandler/handlerequests/v1/requestshandlerutils.go
@@ -15,6 +15,7 @@ import (
 	"github.com/kubescape/kubescape/v2/core/cautils"
 	"github.com/kubescape/kubescape/v2/core/cautils/getter"
 	"github.com/kubescape/kubescape/v2/core/core"
+	"github.com/kubescape/kubescape/v2/httphandler/storage"
 	utilsapisv1 "github.com/kubescape/opa-utils/httpserver/apis/v1"
 	utilsmetav1 "github.com/kubescape/opa-utils/httpserver/meta/v1"
 	reporthandlingv2 "github.com/kubescape/opa-utils/reporthandling/v2"
@@ -79,6 +80,17 @@ func scan(ctx context.Context, scanInfo *cautils.ScanInfo, scanID string) (*repo
 	if err := result.HandleResults(ctx); err != nil {
 		return nil, err
 	}
+	storage := storage.GetStorage()
+	if storage != nil {
+		pr := result.GetResults()
+
+		if err := storage.StorePostureReportResults(ctx, pr); err != nil {
+			return nil, err
+		}
+	} else {
+		logger.L().Info("storage is not initialized")
+	}
+
 	return nil, nil
 }
